Made significant changes to allow compilation (not linking) of EpochCompiler. Still needs EpochVM in order to link. EpochLexer, Tools, and Library also need work.

+++ Shared/Utility/Types/IntegerTypes.h
@@ -5,7 +5,7 @@
 // Platform-specific definitions for integer type widths
 //
 
-typedef __int64 Integer64;
+typedef long long Integer64;
 
 typedef int Integer32;
 typedef unsigned UInteger32;
